+++ client/src/js/helpers/request.js
@@ -1,8 +1,7 @@
-export const URL = `https://web-application-grading.herokuapp.com/`;
+export const URL = `http://localhost:8080/`;
 
 export const getRequest = async (url, options={method: 'GET'}) => {
   const answer = await fetch(url, options);
-  console.log(answer);
   return answer.json();
 };
 export const postRequest = async (url, options) => {
