+++ client/webpack.config.js
@@ -22,7 +22,7 @@ module.exports = {
   devServer: {
     contentBase: path.join(__dirname, "dist"),
     compress: true,
-    port: 5246,
+    port: 4200,
   },
   plugins: [
     new HtmlWebpackPlugin({
